Add tests, use env vars for db creds, gitignore
The settings.py file looks for the environmental variables DB_NAME,
DB_USER, DB_PASSWORD to grab the database credentials. I prefer this
approach to the previous method of adding credentials to the MySQL
config file. If we do want to use a file I propose using a separate
gitignored file within the repo rather than the systemwide MySQL
config file.

The first tests are written for the methods on NewsPost.

I updated the gitignore file with the boilerplate file offered by
Github.