Removed Author model, rebuilding test cases.
I don't think the Author model is necessary anymore with the
new Profile model. We can either add an attribute to the Profile
that specifies whether that profile's user is an author, or we can
use the built-in permission "is_staff" on the User model to
identify authors. I would lean towards the latter.

Started building tests again. The testing goat is angry at me.
Most existing model behavior should be covered, next I will
start writing tests for the views.